Move category reconciliation to ImageView
